// PortalAutoservicio/src/app/models/servidor-publico-model.ts
export class ServidorPublicoModel {
	//Datos generales
	idOrdinal: number;
	claveServidorPublico: string;
	nombre: string;
	primerApellido: string;
	segundoApellido: string;
	genero: string;
	curp: string;
	rfc: string;
	fechaNacimiento: string;
	telefono: string;
	mail: string;
	idSexo: number;
	//Direccion personal
	idTipoDireccionPersonal: number;
	tipoDireccionPersonal: string;
	callePersonal: string;
	idEstadoPersonal: string;
	estadoPersonal: string;
	idMunicipioPersonal: string;
	municipioPersonal: string;
	idColoniaPersonal: string;
	coloniaPersonal: string;
	numeroExteriorPersonal: string;
	numeroInteriorPersonal: string;
	codigoPostalPersonal: string;
	//Direccion fiscal
	idTipoDireccionFiscal: number;
	tipoDireccionFiscal: string;
	calleFiscal: string;
	idEstadoFiscal: string;
	estadoFiscal: string;
	idMunicipioFiscal: string;
	municipioFiscal: string;
	idColoniaFiscal: string;
	coloniaFiscal: string;
	numeroExteriorFiscal: string;
	numeroInteriorFiscal: string;
	codigoPostalFiscal: string;
	//Datos laborales
	nombrePlaza: string;
	idNivelEstudios: string;
	idEstadoCivil: string;
	idPlaza: string;
	issemym: string;
	orden: number;
}